third_party/WebKit/LayoutTests/fast/canvas-api/CanvasRendering2D-prototype-chain.html

<!doctype html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script>

test(function() {
    assert_equals(CanvasRenderingContext2D.prototype.__proto__, Object.prototype);
}, 'Constructor prototype');

test(function() {
    var canvasElement = document.createElement('canvas');
    var context = canvasElement.getContext('2d');
    assert_equals(context.__proto__, CanvasRenderingContext2D.prototype);
}, 'Instance prototype chain');

</script>
